// src/enums/types.enum.ts
export enum InvestorType {
    INDIVIDUAL = "Individual Investor",
    FIRM = "Investment Firm"
}

export enum PromptMessage {
    SUCCESS = "success",
    WARNING = "warning",
    ERROR = "error"
}

export enum FileType {
    IMAGE = "image",
    PDF = "pdf"
}

export enum ApplicationStatus {
    IN_PROGRESS = "inprogress",
    APPROVED = "approved",
    SUBMITTED = "submitted",
    OPENED = "opened",
    REOPENED = "reopened",
    VERIFIED = "verified",
    REJECTED = "rejected",
}